backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai